Created the command line interface again.
You can ask for help or the rate limit. You can tell it to clear the
data, the pull users from a JSON file, and what format to print in
(readable/json). You can also tack on users at the end of the command to
specify who you want that may not be in your file.

All of this is optional and any commands can be ran with each other
(with the exception of rate-limit and help - those will just print some
info and close).

Also made the falsey variable private to hide its existence...